src/requirements.txt
yagmail[all]
pyzbar
python3-picamera2
python-dotenv
websockets
RPi